+++ lib/src/test/java/net/amcintosh/freshbooks/models/PagesTest.java
@@ -1,6 +1,6 @@
 package net.amcintosh.freshbooks.models;
-import static org.junit.Assert.assertEquals;
-import org.junit.Test;
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import org.junit.jupiter.api.Test;
 
 public class PagesTest {