+++ lib/src/test/java/net/amcintosh/freshbooks/models/builders/PaginationQueryBuilderTest.java
@@ -1,9 +1,9 @@
 package net.amcintosh.freshbooks.models.builders;
 
 import net.amcintosh.freshbooks.resources.api.ResourceType;
-import org.junit.Test;
+import org.junit.jupiter.api.Test;
 
-import static org.junit.Assert.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertEquals;
 
 public class PaginationQueryBuilderTest {