+++ lib/src/test/java/net/amcintosh/freshbooks/resources/ClientsTest.java
@@ -10,14 +10,14 @@ import net.amcintosh.freshbooks.TestUtil;
 import net.amcintosh.freshbooks.models.Client;
 import net.amcintosh.freshbooks.models.ClientList;
 import net.amcintosh.freshbooks.models.VisState;
-import org.junit.Test;
+import org.junit.jupiter.api.Test;
 
 import java.io.IOException;
 import java.time.ZoneId;
 import java.time.ZonedDateTime;
 import java.util.Map;
 
-import static org.junit.Assert.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;