+++ lib/src/test/java/net/amcintosh/freshbooks/resources/ItemsTest.java
@@ -9,7 +9,7 @@ import net.amcintosh.freshbooks.TestUtil;
 import net.amcintosh.freshbooks.models.Item;
 import net.amcintosh.freshbooks.models.ItemList;
 import net.amcintosh.freshbooks.models.VisState;
-import org.junit.Test;
+import org.junit.jupiter.api.Test;
 
 import java.io.IOException;
 import java.math.BigDecimal;
@@ -17,7 +17,7 @@ import java.time.ZoneId;
 import java.time.ZonedDateTime;
 import java.util.Map;
 
-import static org.junit.Assert.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;