+++ lib/src/test/java/net/amcintosh/freshbooks/resources/ServicesTest.java
@@ -6,14 +6,16 @@ import com.google.common.collect.ImmutableMap;
 import net.amcintosh.freshbooks.FreshBooksClient;
 import net.amcintosh.freshbooks.FreshBooksException;
 import net.amcintosh.freshbooks.TestUtil;
-import net.amcintosh.freshbooks.models.*;
-import org.junit.Test;
+import net.amcintosh.freshbooks.models.Service;
+import net.amcintosh.freshbooks.models.ServiceList;
+import net.amcintosh.freshbooks.models.VisState;
+import org.junit.jupiter.api.Test;
 
 import java.io.IOException;
 import java.util.Map;
 
-import static org.junit.Assert.assertEquals;
-import static org.junit.Assert.assertTrue;
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertTrue;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;