+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+## 0.6.0
+
 - Added Profit & Loss accounting report
 - Made taxNumber1, taxNumber2 from invoice LineItem readonly due to FreshBooks API change
 - Upgrade dependencies (including JUnit4->5)